requirements.txt
# to install requirements, run ' pip install -r requirements.txt '
# to create your very own requirements file, run ' pip freeze > requirements.txt '
# also check the files that you require before actually installing
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
alpha-vantage==2.3.1
altair==5.2.0
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.9.1.post1
asgiref==3.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
autopep8==1.6.0
Babel==2.14.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
blis==0.7.11
boto3==1.34.49
botocore==1.34.49
CacheControl==0.12.14
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.4
confluent-kafka==2.3.0
coverage==7.4.3
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
distlib==0.3.8
distro==1.9.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.3.0
docopt==0.6.2
editdistance==0.8.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
fbmessenger==6.0.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
fuzzy_matcher==0.1.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google==3.0.0
google-api-core==1.34.1
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-cloud-language==2.13.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
graphql-core==3.2.3
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.48.2
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
jaraco.context==4.3.0
jax==0.4.24
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.3
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
libcst==1.2.0
locket==1.0.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mattermostwrapper==2.2
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==5.2.0
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==2.6.3
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==2.2.1
partd==1.4.1
peewee==3.17.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.28
proto-plus==1.23.0
protobuf==4.23.3
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.17.2
PyJWT==2.3.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.1
pyreadline3==3.4.1
pyrsistent==0.20.0
pytest==8.0.2
pytest-cov==4.1.0
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa==3.6.19
rasa-sdk==3.6.2
rasa_addons==0.9.29.3
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-futures==1.0.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
safetensors==0.4.2
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.12.0
sentry-sdk==1.14.0
sgqlc==16.3
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.27.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.51
sqlparse==0.4.2
srsly==2.4.8
starlette==0.36.3
streamlit==1.31.1
structlog==23.3.0
structlog-sentry==2.0.3
sympy==1.12
tabulate==0.9.0
tarsafe==0.0.4
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
twilio==8.2.2
typer==0.9.0
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
validators==0.22.0
virtualenv==20.25.1
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.1
wikipedia==1.4.0
Wikipedia-API==0.6.0
wolframalpha==5.0.0
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.4
yfinance==0.2.37
zipp==3.17.0
